Refactor Typography component to use Static themeing
[#9]

* Refactored Typography component to use static themeing
