+++ packages/common/src/components/Typography/Typography.module.scss
@@ -1,0 +1,64 @@
+.h1 {
+  text-align: left;
+  font: normal normal bold 48px/62px "PT Sans";
+  letter-spacing: 0px;
+  color: var(--typography-color);
+  opacity: 1;
+}
+
+.h2 {
+  text-align: left;
+  font: normal normal bold 36px/47px "PT Sans";
+  letter-spacing: 0px;
+  color: var(--typography-color);
+  opacity: 1;
+}
+
+.h3 {
+  text-align: left;
+  font: normal normal bold 24px/33px "PT Sans";
+  letter-spacing: 0px;
+  color: var(--typography-color);
+  opacity: 1;
+}
+
+.h4 {
+  text-align: left;
+  font: normal normal bold 18px/24px "Open Sans";
+  letter-spacing: 0px;
+  color: var(--typography-color);
+  opacity: 1;
+}
+
+.h5 {
+  text-align: left;
+  font: normal normal bold 16px/22px "Open Sans";
+  letter-spacing: 0px;
+  color: var(--typography-color);
+  opacity: 1;
+}
+
+.h6 {
+  text-align: left;
+  font: normal normal bold 14px/19px "Open Sans";
+  letter-spacing: 0px;
+  color: var(--typography-color);
+  opacity: 1;
+}
+
+.p {
+  text-align: left;
+  font: normal normal normal 16px/22px "Open Sans";
+  font-weight: 500;
+  letter-spacing: 0px;
+  color: var(--typography-color);
+  opacity: 1;
+}
+
+.link {
+  text-align: left;
+  font: normal normal normal 16px/22px "Open Sans";
+  letter-spacing: 0px;
+  color: #0085c9;
+  opacity: 1;
+}
