+++ packages/common/src/components/Typography/constants.ts
@@ -1,0 +1,10 @@
+export enum TypographyVariant_ENUM {
+  h1 = "h1",
+  h2 = "h2",
+  h3 = "h3",
+  h4 = "h4",
+  h5 = "h5",
+  h6 = "h6",
+  p = "p",
+  link = "link",
+}
