+++ packages/common/src/components/Typography/index.ts
@@ -1,2 +1,4 @@
-export { Typography, EnumTypographyVariant } from "./Typography";
+export { default as Typography } from "./Typography";
 export type { TypeTypographyProps } from "./Typography";
+
+export { TypographyVariant_ENUM } from "./constants";
