+++ packages/common/src/components/Typography/storybook.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Typography, EnumTypographyVariant } from "./Typography";
+import { Typography, TypographyVariant_ENUM } from "../Typography";
 // This default export determines where you story goes in the story list
 export default {
   title: "Typography",
@@ -8,21 +8,21 @@ export default {
 
 const TypographyStory = () => (
   <>
-    <Typography variant={EnumTypographyVariant.h1}>Header H1</Typography>
+    <Typography variant={TypographyVariant_ENUM.h1}>Header H1</Typography>
     <br />
-    <Typography variant={EnumTypographyVariant.h2}>Header H2</Typography>
+    <Typography variant={TypographyVariant_ENUM.h2}>Header H2</Typography>
     <br />
-    <Typography variant={EnumTypographyVariant.h3}>Header H#</Typography>
+    <Typography variant={TypographyVariant_ENUM.h3}>Header H#</Typography>
     <br />
-    <Typography variant={EnumTypographyVariant.h4}>Header H4</Typography>
+    <Typography variant={TypographyVariant_ENUM.h4}>Header H4</Typography>
     <br />
-    <Typography variant={EnumTypographyVariant.h5}>Header H5</Typography>
+    <Typography variant={TypographyVariant_ENUM.h5}>Header H5</Typography>
     <br />
-    <Typography variant={EnumTypographyVariant.h6}>Header H6</Typography>
+    <Typography variant={TypographyVariant_ENUM.h6}>Header H6</Typography>
     <br />
-    <Typography variant={EnumTypographyVariant.p}>Text p</Typography>
+    <Typography variant={TypographyVariant_ENUM.p}>Text p</Typography>
     <br />
-    <Typography variant={EnumTypographyVariant.link}>Link</Typography>
+    <Typography variant={TypographyVariant_ENUM.link}>Link</Typography>
     <br />
   </>
 );
